requirements.txt
PyYAML==3.11
requests==2.4.3
